Create seperate header file for NetworkConnection and network states
