+++ src/include/network/network_thread.h
@@ -32,7 +32,6 @@
 #include "common/logger.h"
 #include "configuration/configuration.h"
 #include "container/lock_free_queue.h"
-#include "network_manager.h"
 #include "protocol_handler.h"
 
 namespace peloton {
